pyproject.toml
[project]
name = "coregenai"
version = "0.1.0"
description = "CoreGenAI is the unified Python SDK for Google's Generative AI that eliminates confusion, handles complexity, and just works."
authors = [{ name = "Saadman Rafat", email = "saadman@outlook.com" }]
keywords = ["gemini", "vertex-ai", "google", "ai", "llm", "generative-ai", "genai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.50.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "vertexai>=1.71.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"